Fix equal button handler
Clear secondOperand and currentOperation, and set firstOperand to the
result. In this case the user when start clicking on digits, the new
digit will be appended to the existing number on the current screen.

If the user clicked an operator instead, the number on the current
screen (the result of previous operation) will be the first operand for
the next operation.